backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
psycopg[binary]==3.2.9
neo4j==5.28.1
sqlalchemy==2.0.41
alembic==1.16.1
email-validator==2.2.0
PyMuPDF==1.26.0
python-multipart==0.0.20
httpx==0.28.1
pytesseract==0.3.13
Pillow==11.2.1
